Added error notification
